Fixes irq_to_hwirq() API should have a return type of uint64 (#428)

+++ platform/pal_baremetal/common/include/pal_common_support.h
@@ -1,5 +1,5 @@
 /** @file
- * Copyright (c) 2020-2023, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2020-2024, Arm Limited or its affiliates. All rights reserved.
  * SPDX-License-Identifier : Apache-2.0
 
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -502,7 +502,7 @@ typedef struct {
   uint32_t  vector_lower_addr; ///< Base Address of the controller
   uint32_t  vector_data;       ///< Base Address of the controller
   uint32_t  vector_control;    ///< IRQ to install an ISR
-  uint32_t  vector_irq_base;   ///< Base IRQ for the vectors in the block
+  uint64_t  vector_irq_base;   ///< Base IRQ for the vectors in the block
   uint32_t  vector_n_irqs;     ///< Number of irq vectors in the block
   uint32_t  vector_mapped_irq_base; ///< Mapped IRQ number base for this MSI
 }PERIPHERAL_VECTOR_BLOCK;
